requirements.txt
asgiref==3.8.1
blinker==1.9.0
Brotli==1.1.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
ConfigArgParse==1.7
coverage==7.6.12
Django==5.1.5
django-stubs==5.1.2
django-stubs-ext==5.1.2
djangorestframework==3.15.2
djangorestframework-stubs==3.15.2
djangorestframework_simplejwt==5.4.0
dnspython==2.7.0
drf-yasg==1.21.8
email-validator==2.2.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
gevent==24.11.1
geventhttpclient==2.3.3
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
locust==2.32.10
MarkupSafe==3.0.2
msgpack==1.1.0
mypy==1.14.1
mypy-extensions==1.0.0
mysqlclient==2.2.7
packaging==24.2
pluggy==1.5.0
psutil==7.0.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytest==8.3.4
pytest-cov==6.0.0
pytest-django==4.10.0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
requests==2.32.3
sentry-sdk==2.22.0
SQLAlchemy==2.0.38
sqlparse==0.5.3
types-PyYAML==6.0.12.20241230
types-requests==2.32.0.20241016
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
zope-event==5.0
zope-interface==7.2